Fixed a weird and irritating bug with the session cookie.
If you go to uio.no/brukerinfo/ you will get redirected to brukerinfo.uio.no//.
A bug in the session cookie parameter would then create a session cookie with
path set to //, and not /. This new cookie would override the logged on session,
but only on the url brukerinfo.uio.no/, so that the site doesn't think you're
logged on anyway. You would then be redirected to the logon page and logged off.

Also changed the session name, to prevent already set session cookies to still
create problems.

Thanks to stalk for finding the bug.