Use injection to assemble features
A new config option (but reasonable defaults) controls features:
an array of modules to have. Each feature will load itself and
its links on the front page, as well as the tabs. Most php pages
are replaced with the injection interface, and use PATH_INFO for
location (i.e. https://brukerinfo/person/password.php is now
https://brukerinfo/index.php/person/password/).