Fix port numbers in UserForward
If the user accesses a specific module, but needs authentication, the
session variable UserForward is set. On successful logon, the user is
redirected to UserForward.

If the BASE_URL contains anything other than scheme and host, the
redirect would not point to the correct location.  This change fixes the
logon redirect by keeping all parameters from url_parse except path,
query and fragment, which should come from the UserForward session
variable.

Issue: WOFH-21