Adjust access to brukerinfo for UiT
Prior to this commit a users at UiT could go to /groups/personal on the
brukerinfo website and create their own personal file group even though
they were not meant to be able to.

This adds a check that redirects any user that does not have access to
create groups back to /groups/ if they try to go to groups/personal.

Issue: CRB-2832