Fix brukerinfo missing affiliations error
Fixed an error when the user did not have any affiliations and the program would throw a bunch of notices regarding missing values that was expected to be put into the template messages.
Added a new message if the case is that the user does not have any affiliations, the string now reads
"No affiliations registred" in English and
"Ingen tilknytninger er registrert" in Norwegian.

Issue: CRB-3059